+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## Change Log
 
 ### current version
-- feature: Charger Mode (/electrical.chargers.<#>.mode) is supported
-- feature: Charger State of Charge is supported (/electrical.chargers.<#>.capacity.stateOfCharge')
+- feature: Charger Mode is supported (electrical.chargers.<#>.mode)
+- feature: Charger State of Charge is supported (electrical.chargers.<#>.capacity.stateOfCharge)
 
 - fix: Cleanup HomeKit device meta data for new devices. Should not have side effects on existing devices.